bot/requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
future==0.16.0
gunicorn==19.8.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
line-bot-sdk
MarkupSafe==1.0
requests==2.18.4
urllib3==1.22
Werkzeug==0.14.1
BeautifulSoup4
bs4